[hw7: типизация] теперь на typescript: server/views
