+++ server/utils.ts
@@ -1,35 +1,12 @@
 const TYPES = ['info', 'critical'];
+import {eventItemType, eventsJsonType, paginationResultType} from './types';
 
-type eventItemType = {
-  type: string,
-  title: string,
-  source: string,
-  time: string,
-  description: string | null,
-  icon: string,
-  size: 's' | 'm' | 'l',
-  data?: {
-    type?: string,
-  }
-}
-
-type eventsJsonType = {
-  events: eventItemType[],
-};
-
-type paginationResultType = {
-  errors: boolean,
-  limit?: number,
-  page?: number,
-};
-
-
-function pad(num: string, length: number = 2, symb: string = '0'): string {
+export function pad(num: string, length: number = 2, symb: string = '0'): string {
   const n = num.toString();
   return n.length >= length ? n : pad(symb + n, length, symb);
 }
 
-function dateDiff(date1: Date, date2: Date):string {
+export function dateDiff(date1: Date, date2: Date):string {
   let diff = +date2 - +date1;
 
   let hours: number = Math.floor(diff / 60 / 60 / 1000);
@@ -45,7 +22,7 @@ function dateDiff(date1: Date, date2: Date):string {
  * limit задает, количество элементов на странице, по умолчанию 0
  * если limit === 0, выведутся все элементы
  */
-function preparePaginationParams(
+export function preparePaginationParams(
   queryLimit: string = '0',
   queryPage: string = '1'
 ): paginationResultType {
@@ -53,6 +30,8 @@ function preparePaginationParams(
   const page: number = +queryPage;
   const result: paginationResultType = {
     errors: false,
+    page: 1,
+    limit: 0,
   };
 
   if (isNaN(limit) || isNaN(page) || limit < 0 || page <= 0) {
@@ -74,7 +53,7 @@ function preparePaginationParams(
   return result;
 }
 
-function filterEvents(query: string, eventsJson: eventsJsonType): eventsJsonType {
+export function filterEvents(query: string, eventsJson: eventsJsonType): eventsJsonType {
   if (!query) return eventsJson;
   const types: string[] = query.split(':');
 
@@ -88,7 +67,7 @@ function filterEvents(query: string, eventsJson: eventsJsonType): eventsJsonType
   }
 }
 
-function pagitateEvents(data: eventsJsonType, limit: number, page: number): eventsJsonType {
+export function pagitateEvents(data: eventsJsonType, limit: number, page: number): eventsJsonType {
   const startIndex: number = (page - 1) * limit;
   const endIndex: number = limit ? startIndex + limit : data.events.length;
   const events: eventItemType[] = data.events.slice(startIndex, endIndex);
@@ -98,16 +77,7 @@ function pagitateEvents(data: eventsJsonType, limit: number, page: number): even
   }
 }
 
-function checkFilter(filters: string): boolean {
+export function checkFilter(filters: string): boolean {
   const filtersArr: string[] = filters.split(':');
   return !filtersArr.some(item => !TYPES.includes(item));
 }
-
-module.exports = {
-  pad,
-  dateDiff,
-  filterEvents,
-  pagitateEvents,
-  preparePaginationParams,
-  checkFilter,
-};
